+++ template/trpc/server/context.ts
@@ -1,0 +1,16 @@
+import type { inferAsyncReturnType } from "@trpc/server";
+import type { createSolidAPIHandlerContext } from "solid-start-trpc";
+
+export const createContextInner = async (
+  opts: createSolidAPIHandlerContext
+) => {
+  return {
+    ...opts,
+  };
+};
+
+export const createContext = async (opts: createSolidAPIHandlerContext) => {
+  return await createContextInner(opts);
+};
+
+export type IContext = inferAsyncReturnType<typeof createContext>;