+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-jd-app
 
+## 0.1.27
+
+### Patch Changes
+
+- [`0d70686`](https://github.com/OrJDev/create-jd-app/commit/0d70686bcc0229fcb5ca55ad11de024524a70cfa) Thanks [@OrJDev](https://github.com/OrJDev)! - Fix: prpc vite / packages
+
+## 0.1.26
+
+### Patch Changes
+
+- [`83d5d20`](https://github.com/OrJDev/create-jd-app/commit/83d5d206d2871a8da201373084ca255cf3f747d9) Thanks [@OrJDev](https://github.com/OrJDev)! - Feat: tRPC installer
+
 ## 0.1.25
 
 ### Patch Changes